Added Hand Result class
This class is used to keep track of the result of each player's hand.

It's needed to be able to determine the winner of each round, which has the type of hand and the cards each player had at the end of the round.

Implementation of getting the winner is not added in this commit.